Refactor and optimize various features and utilities
- Updated drawing logic in Outlines.cpp for improved rectangle rendering.
- Enhanced Paint.cpp to prevent iterator invalidation during map modifications.
- Cleaned up ProjectileSim.cpp by removing redundant return statements.
- Fixed typo in SpyWarning.cpp configuration checks for visuals.
- Added original angles storage in AutoAirblast.cpp for better aim assist.
- Corrected configuration variable names in AutoBackstab.cpp for consistency.
- Improved AutoVaccinator.cpp to safely handle null pointers when accessing weapon entities.
- Added health color handling in VisualUtils.cpp to prevent division by zero.
- Adjusted CL_Move.cpp to ensure rapid fire ticks do not exceed available ticks.
- Implemented visual interpolation tracking in IBaseClientDLL_FrameStageNotify.cpp for smoother player rendering.
- Updated Draw.cpp to use safer string formatting functions.
- Fixed entity cache update logic in Entities.cpp to include the highest entity index.
- Enhanced SDK.h to include safety checks for ConVar pointers.
- Improved error handling in Config.h for JSON parsing.
- Refined HookManager.cpp to assert hook initialization and enabling.
- Updated InterfaceManager.cpp to dereference interface pointers safely.
- Enhanced Math.h to use atan2 for angle calculations.
- Added signature validation in Memory.cpp to prevent out-of-bounds access.
- Improved SignatureManager.cpp to handle uninitialized signatures gracefully.
- Enhanced Utils.h to ensure safe string conversions and random number generation.

+++ SEOwnedDE/SEOwnedDE/src/Utils/Color/Color.h
@@ -32,7 +32,7 @@ namespace ColorUtils
 
 	inline unsigned long ToDWORD(Color_t x)
 	{
-		return (x.r & 0xFF) << 24 | (x.g & 0xFF) << 16 | (x.b & 0xFF) << 8 | x.a & 0xFF;
+		return static_cast<unsigned long>(x.r & 0xFF) << 24 | static_cast<unsigned long>(x.g & 0xFF) << 16 | static_cast<unsigned long>(x.b & 0xFF) << 8 | static_cast<unsigned long>(x.a & 0xFF);
 	}
 
 	inline Color_t Rainbow(float offset, float rate = 1.f)
